ug-admin/backend/pyproject.toml
[tool.poetry]
name = "ug-admin-backend"
version = "0.1.0"
description = "Undergraduation.com Admin Dashboard Backend"
authors = ["Undergraduation Team"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic-settings = "^2.0.3"
firebase-admin = "^6.2.0"
google-cloud-firestore = "^2.13.1"
python-dotenv = "^1.0.0"
pydantic = {extras = ["email"], version = "^2.11.9"}
python-multipart = "^0.0.6"
jinja2 = "^3.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
httpx = "^0.25.2"
pytest-asyncio = "^0.21.1"
ruff = "^0.1.6"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.1"
types-requests = "^2.31.0.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
